+++ nautilus_trader/common/logging.pyx
@@ -13,12 +13,11 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+import asyncio
 import logging
-import multiprocessing
 import os
 import platform
 from platform import python_version
-import queue
 import sys
 import threading
 import traceback
@@ -38,6 +37,7 @@ from nautilus_trader.common.clock cimport LiveClock
 from nautilus_trader.common.logging cimport LogLevel
 from nautilus_trader.common.logging cimport LogMessage
 from nautilus_trader.common.logging cimport Logger
+from nautilus_trader.common.queue cimport Queue
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.datetime cimport format_iso8601_us
 
@@ -77,8 +77,6 @@ cdef class LogLevelParser:
             return "CRT"
         elif value == 7:
             return "FTL"
-        else:
-            return "UNDEFINED"
 
     @staticmethod
     cdef LogLevel from_str(str value):
@@ -96,8 +94,6 @@ cdef class LogLevelParser:
             return LogLevel.CRITICAL
         elif value == "FTL":
             return LogLevel.FATAL
-        else:
-            return LogLevel.UNDEFINED
 
     @staticmethod
     def to_str_py(int value):
@@ -143,18 +139,7 @@ cdef class LogMessage:
         self.text = text
         self.thread_id = thread_id
 
-    cdef str level_string(self):
-        """
-        Return the string representation of the log level.
-
-        Returns
-        -------
-        str
-
-        """
-        return LogLevelParser.to_str(self.level)
-
-    cdef str as_string(self):
+    cdef inline str as_string(self):
         """
         Return the string representation of the log message.
 
@@ -184,7 +169,7 @@ cdef class Logger:
         bint console_prints=True,
         bint log_thread=False,
         bint log_to_file=False,
-        str log_file_path not None="",
+        str log_file_dir not None="",
     ):
         """
         Initialize a new instance of the `Logger` class.
@@ -209,15 +194,15 @@ cdef class Logger:
             If log messages should include the thread.
         log_to_file : bool
             If log messages should be written to the log file.
-        log_file_path : str
-            The name of the log file (cannot be None if log_to_file is True).
+        log_file_dir : str
+            The log file directory.
 
         Raises
         ------
         ValueError
             If name is not a valid string.
         ValueError
-            If log_file_path is not a valid string.
+            If log_file_dir is not a valid string.
 
         """
         if name is not None:
@@ -225,9 +210,9 @@ cdef class Logger:
         else:
             name = "tmp"
         if log_to_file:
-            if log_file_path == "":
-                log_file_path = "log/"
-            Condition.valid_string(log_file_path, "log_file_path")
+            if log_file_dir == "":
+                log_file_dir = "log/"
+            Condition.valid_string(log_file_dir, "log_file_dir")
 
         self.name = name
         self.bypass_logging = bypass_logging
@@ -238,19 +223,52 @@ cdef class Logger:
         self._console_prints = console_prints
         self._log_thread = log_thread
         self._log_to_file = log_to_file
-        self._log_file_path = log_file_path
-        self._log_file = f"{self._log_file_path}{self.name}-{self.clock.utc_now().date().isoformat()}.log"
+        self._log_file_dir = log_file_dir
+        self._log_file_path = f"{self._log_file_dir}{self.name}-{self.clock.utc_now().date().isoformat()}.log"
         self._log_store = []
         self._logger = logging.getLogger(name)
         self._logger.setLevel(logging.DEBUG)
 
         # Setup log file handling
         if log_to_file:
-            if not os.path.exists(log_file_path):
+            if not os.path.exists(self._log_file_dir):
                 # Create directory if it does not exist
-                os.makedirs(log_file_path)
-            self._log_file_handler = logging.FileHandler(self._log_file)
+                os.makedirs(self._log_file_dir)
+            self._log_file_handler = logging.FileHandler(self._log_file_path)
             self._logger.addHandler(self._log_file_handler)
+
+    cpdef str get_log_file_dir(self):
+        """
+        Return the current log file directory.
+
+        Returns
+        -------
+        str
+
+        """
+        return self._log_file_dir
+
+    cpdef str get_log_file_path(self):
+        """
+        Return the current log file path.
+
+        Returns
+        -------
+        str
+
+        """
+        return self._log_file_path
+
+    cpdef list get_log_store(self):
+        """
+        Return the log store of message strings.
+
+        Returns
+        -------
+        list[str]
+
+        """
+        return self._log_store
 
     cpdef void change_log_file_name(self, str name) except *:
         """
@@ -264,25 +282,14 @@ cdef class Logger:
         """
         Condition.valid_string(name, "name")
 
-        self._log_file = f"{self._log_file_path}{name}.log"
+        self._log_file_path = f"{self._log_file_dir}{name}.log"
         self._logger.removeHandler(self._log_file_handler)
-        self._log_file_handler = logging.FileHandler(self._log_file)
+        self._log_file_handler = logging.FileHandler(self._log_file_path)
         self._logger.addHandler(self._log_file_handler)
 
     cpdef void log(self, LogMessage message) except *:
         """Abstract method (implement in subclass)."""
         raise NotImplementedError("method must be implemented in the subclass")
-
-    cpdef list get_log_store(self):
-        """
-        Return the log store of message strings.
-
-        Returns
-        -------
-        list[str]
-
-        """
-        return self._log_store
 
     cpdef void clear_log_store(self) except *:
         """
@@ -492,14 +499,17 @@ cdef class LoggerAdapter:
         LogColor color,
         str message,
     ) except *:
+        cdef:
+            LogMessage msg
         if not self.is_bypassed:
-            self._logger.log(LogMessage(
+            msg = LogMessage(
                 self._logger.clock.utc_now(),
                 level,
                 color,
                 self._format_message(message),
-                threading.current_thread().ident),
+                threading.current_thread().ident,
             )
+            self._logger.log(msg)
 
     cdef inline str _format_message(self, str message):
         # Add the components name to the front of the log message
@@ -584,7 +594,7 @@ cdef class TestLogger(Logger):
         bint console_prints=True,
         bint log_thread=False,
         bint log_to_file=False,
-        str log_file_path not None="log/",
+        str log_file_dir not None="log/",
     ):
         """
         Initialize a new instance of the `TestLogger` class.
@@ -609,19 +619,19 @@ cdef class TestLogger(Logger):
             If log messages should include the thread.
         log_to_file : bool
             If log messages should write to the log file.
-        log_file_path : str
-            The name of the log file (cannot be None if log_to_file is True).
+        log_file_dir : str
+            The log file directory (cannot be empty if log_to_file is True).
 
         Raises
         ------
         ValueError
             If name is not a valid string.
         ValueError
-            If log_file_path is not a valid string.
+            If log_file_dir is not a valid string.
 
         """
-        if log_file_path is "":
-            log_file_path = "log/"
+        if log_file_dir is "":
+            log_file_dir = "log/"
         super().__init__(
             clock,
             name,
@@ -632,7 +642,7 @@ cdef class TestLogger(Logger):
             console_prints,
             log_thread,
             log_to_file,
-            log_file_path,
+            log_file_dir,
         )
 
     cpdef void log(self, LogMessage message) except *:
@@ -652,12 +662,12 @@ cdef class TestLogger(Logger):
 
 cdef class LiveLogger(Logger):
     """
-    Provides a high-performance logger which runs in a separate process for live
-    operations.
+    Provides a high-performance logger which runs on the event loop.
     """
 
     def __init__(
         self,
+        loop not None,
         LiveClock clock not None,
         str name=None,
         bint bypass_logging=False,
@@ -668,13 +678,16 @@ cdef class LiveLogger(Logger):
         bint console_prints=True,
         bint log_thread=False,
         bint log_to_file=False,
-        str log_file_path not None="logs/",
+        str log_file_dir not None="logs/",
+        int maxsize=10000,
     ):
         """
         Initialize a new instance of the `LiveLogger` class.
 
         Parameters
         ----------
+        loop : asyncio.AbstractEventLoop
+            The event loop to run the logger on.
         clock : LiveClock
             The clock for the logger.
         name : str
@@ -693,15 +706,17 @@ cdef class LiveLogger(Logger):
             If log messages should include the thread.
         log_to_file : bool
             If log messages should write to the log file.
-        log_file_path : str
-            The name of the log file (cannot be None if log_to_file is True).
+        log_file_dir : str
+            The log file directory (cannot be empty if log_to_file is True).
+        maxsize : int, optional
+            The maximum capacity for the log queue.
 
         Raises
         ------
         ValueError
             If the name is not a valid string.
         ValueError
-            If the log_file_path is not a valid string.
+            If the log_file_dir is not a valid string.
 
         """
         super().__init__(
@@ -714,17 +729,14 @@ cdef class LiveLogger(Logger):
             console_prints,
             log_thread,
             log_to_file,
-            log_file_path,
+            log_file_dir,
         )
 
-        if run_in_process:
-            self._queue = multiprocessing.Queue(maxsize=10000)
-            self._process = multiprocessing.Process(target=self._consume_messages, daemon=True)
-            self._process.start()
-        else:
-            self._queue = queue.Queue(maxsize=10000)
-            self._thread = threading.Thread(target=self._consume_messages, daemon=True)
-            self._thread.start()
+        self._loop = loop
+        self._queue = Queue(maxsize=maxsize)
+
+        self._run_task = None
+        self.is_running = False
 
     cpdef void log(self, LogMessage message) except *:
         """
@@ -732,6 +744,9 @@ cdef class LiveLogger(Logger):
 
         If the internal queue is already full then will log a warning and block
         until queue size reduces.
+
+        If the event loop is not running then messages will be passed directly
+        to the `Logger` base class for logging.
 
         Parameters
         ----------
@@ -741,38 +756,51 @@ cdef class LiveLogger(Logger):
         """
         Condition.not_none(message, "message")
 
-        try:
-            self._queue.put_nowait(message)
-        except queue.Full:
-            queue_full_msg = LogMessage(
-                timestamp=self.clock.utc_now(),
-                level=LogLevel.WARNING,
-                text=f"LiveLogger: Blocking on `put` as queue full at {self._queue.qsize()} items.",
-                thread_id=threading.current_thread().ident,
-            )
-            self._queue.put(queue_full_msg)
-            self._queue.put(message)  # Block until qsize reduces below maxsize
+        if self.is_running:
+            try:
+                self._queue.put_nowait(message)
+            except asyncio.QueueFull:
+                queue_full_msg = LogMessage(
+                    timestamp=self.clock.utc_now(),
+                    level=LogLevel.WARNING,
+                    color=LogColor.YELLOW,
+                    text=f"LiveLogger: Blocking on `_queue.put` as queue full at {self._queue.qsize()} items.",
+                    thread_id=threading.current_thread().ident,
+                )
+                self._log(queue_full_msg)
+                self._loop.create_task(self._queue.put(message))  # Blocking until qsize reduces
+        else:
+            # If event loop is not running then pass message directly to the
+            # base class to log.
+            self._log(message)
+
+    cpdef void start(self) except *:
+        """
+        Start the logger on a running event loop.
+        """
+        if not self.is_running:
+            self._run_task = self._loop.create_task(self._consume_messages())
+        self.is_running = True
 
     cpdef void stop(self) except *:
-        self._queue.put_nowait(None)  # Sentinel message pattern
+        """
+        Stop the logger by cancelling the internal event loop task.
 
-    cpdef void _consume_messages(self) except *:
-        cdef LogMessage message
+        Future messages sent to the logger will be passed directly to the
+        `Logger` base class for logging.
+
+        """
+        if self._run_task:
+            self._run_task.cancel()
+        self.is_running = False
+
+    async def _consume_messages(self):
         try:
             while True:
-                message = self._queue.get()
-                if message is None:  # Sentinel message (fast c-level check)
-                    break
-                self._log(message)
-        except KeyboardInterrupt:
-            if self._process:
-                # Logger is running in a separate process.
-                # Here we have caught a single SIGTERM / keyboard interrupt from
-                # the main thead, this is to allow final log messages to be
-                # processed. Because daemon=True is set, when the main
-                # thread exits this will then terminate the process.
-                while True:
-                    message = self._queue.get()
-                    if message is None:
-                        break
-                    self._log(message)
+                self._log(await self._queue.get())
+        except asyncio.CancelledError:
+            pass
+        finally:
+            # Pass remaining messages directly to the base class
+            while not self._queue.empty():
+                self._log(self._queue.get_nowait())