+++ tests/unit_tests/execution/test_execution_engine.py
@@ -251,7 +251,7 @@ class ExecutionEngineTests(unittest.TestCase):
             Quantity("1.00000000"),
         )
 
-        order.apply(event=TestStubs.event_order_submitted(order))
+        order.apply(TestStubs.event_order_submitted(order))
 
         fill1 = TestStubs.event_order_filled(
             order,
@@ -261,7 +261,7 @@ class ExecutionEngineTests(unittest.TestCase):
             last_px=Price("50000.00000000"),
         )
 
-        order.apply(event=fill1)
+        order.apply(fill1)
         position = Position(fill=fill1)
 
         self.database.add_order(order)